+++ app/src/js/App.js
@@ -26,7 +26,6 @@ import Sections from './components/Sections';
 import Questions from './components/Questions';
 import Workflows from './components/Workflows';
 import Metrics from './components/Metrics';
-import MyMetrics from './components/MyMetrics';
 import Roles from './components/Roles';
 import Conversations from './components/Conversations';
 import Rules from './components/Rules';
@@ -69,7 +68,6 @@ const MainRoutes = ({ activeRoute }) => {
     { path: '/groups', component: Groups },
     { path: '/workflows', component: Workflows },
     { path: '/metrics', component: Metrics },
-    { path: '/mymetrics', component: MyMetrics },
     { path: '/roles', component: Roles },
     { path: '/conversations/', component: Conversations },
     { path: '/rules', component: Rules },