+++ app/src/js/paths/metrics.js
@@ -7,8 +7,9 @@ const routes = [
 const requestRoutes = [
   ['', '', null, ''],
   [strings.all_metrics, '/metrics/', null, 'sidebar__nav--back'],
-  ["DAAC Metrics", '/metrics/daacs', null, 'sidebar__nav--back'],
-  ["User Metrics", '/metrics/users', null, 'sidebar__nav--back'],
+  ['DAAC Metrics', '/metrics/daacs', null, 'sidebar__nav--back', 'METRICS', 'READ'],
+  ['User Metrics', '/metrics/users', null, 'sidebar__nav--back', 'METRICS', 'READ'],
+  ['My Metrics', '/metrics/mymetrics', null, 'sidebar__nav--back'],
 ];
 
 const singleRoutes = [
@@ -30,7 +31,8 @@ const handler = {
           return d;
         }
         return empty;
-    })} else {
+      });
+    } else {
       return routes;
     }
   }